Make functions return empty tuple by default
Currently, functions that do not return anything are a special case from the perspective of the type system. There is no way to capture anything from a call to such functions. In fact, a call to a function that does not return anything is treated as a semantic error.

E.g.

```
contract Foo:

def child():
    y = 1

def parent():
    x = self.child()  # semantic error
```

We [recently discussed](https://github.com/ethereum/fe/pull/103#discussion_r514078077) to change this and follow the model of rust which treats any function that does not explicitly return a value as one that returns an empty tuple `()`.

The core motivation for that is that it generalizes the handling of functions by getting rid of different branching logic based on whether a function does or does not return a value. In the new models, functions simply always return, even if only a zero sized empty tuple.

This means, that the code above would be valid and that the type of `x` would be `()` since this is what `child` returns.

These function definitions should be equivalent and the shorter ones can be treated as syntactic sugar to the first.

Full length form:
```
def child() -> ():
    y = 1
    return ()
```

No explicit return in signature:
```
def child():
    y = 1
    return ()
```

No explicit return statement:
```
def child() -> ():
    y = 1
```

No return in signature and function body:
```
def child():
    y = 1
```

**Notice that support for the `()` syntax is out of scope for this PR and is only meant to explain the planned changes that this PR is working towards.**